# api_service/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.Cookie:
    properties:
      domain:
        type: string
      expires:
        description: Use pointer to handle null/zero time
        type: string
      http_only:
        type: boolean
      id:
        type: integer
      is_sharable:
        type: boolean
      last_updated_from_extension_at:
        type: string
      name:
        type: string
      path:
        type: string
      same_site:
        type: string
      secure:
        type: boolean
      user_id:
        type: integer
      value:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Cookie Syncer browser extension.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Cookie Syncer API
  version: "1.0"
paths:
...
  /auth/test:
    get:
      description: A simple endpoint to check if the provided API key in the `x-api-key`
        header is valid and associated with a user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  properties:
                    role:
                      type: string
                    user_id:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Test API Key
      tags:
      - Auth
  /cookies/{domain}:
    get:
      description: Retrieves cookies for a specific domain. By default, returns an
        HTTP header string. Use ?format=json to get structured JSON.
      parameters:
      - description: Domain
        in: path
        name: domain
        required: true
        type: string
      - description: Output format
        enum:
        - json
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get cookies for a domain
      tags:
      - Cookies
  /cookies/{domain}/{name}:
    get:
      description: Retrieves the raw value of a specific cookie, returned in the 'data'
        field.
      parameters:
      - description: Domain
        in: path
        name: domain
        required: true
        type: string
      - description: Cookie Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a single cookie's value
      tags:
      - Cookies
  /cookies/all:
    get:
      description: Retrieves all cookies for the authenticated user. By default, groups
        them by domain and returns them as HTTP header strings. Use ?format=json to
        get structured JSON.
      parameters:
      - description: Output format
        enum:
        - json
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all cookies
      tags:
      - Cookies
  /pool/cookies/{domain}:
    get:
      description: |-
        Retrieves all sharable cookies for a given domain from users who have opted into sharing.
        This endpoint is protected by a dedicated Pool Access Key (`x-pool-key` header), not a user's API key.
        By default, returns an array of strings, where each string is a user's cookies formatted as an HTTP 'Cookie' header.
        Use `?format=json` to get a structured JSON response, where each element contains the user's ID and their list of cookies.
      parameters:
      - description: The domain to fetch cookies for
        in: path
        name: domain
        required: true
        type: string
      - description: Output format
        enum:
        - json
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JSON response with `?format=json`
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    properties:
                      cookies:
                        items:
                          $ref: '#/definitions/model.Cookie'
                        type: array
                      user_id:
                        type: integer
                    type: object
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - PoolKeyAuth: []
      summary: Get sharable cookies by domain
      tags:
      - Pool
...
  /user/settings:
    get:
      description: Retrieves settings for the authenticated user, such as whether
        cookie sharing is enabled.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  properties:
                    sharing_enabled:
                      type: boolean
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user settings
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates settings for the authenticated user, such as enabling or
        disabling cookie sharing.
      parameters:
      - description: Settings payload
        in: body
        name: body
        required: true
        schema:
          properties:
            sharing_enabled:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user settings
      tags:
      - User
securityDefinitions:
  AdminKeyAuth:
    in: header
    name: x-admin-key
    type: apiKey
  ApiKeyAuth:
    in: header
    name: x-api-key
    type: apiKey
  PoolKeyAuth:
    in: header
    name: x-pool-key
    type: apiKey
swagger: "2.0"